[CP-3816] `PrivacyPolicyFlow` was extracted from `CheckInitRequirements` and moved to a standalone file
